Fixed all the warnings that could be fixed again.

+++ SDL_Game_Framework/AW_Sprite_Interface.cpp
@@ -35,7 +35,7 @@ bool AW_Sprite_Interface::Move_By(int xAmount, int yAmount)
 	return false;
 }
 
-bool AW_Sprite_Interface::Move_To(int x, int y)
+bool AW_Sprite_Interface::Move_To(float x, float y)
 {
 	set_world_position(x, y);
 	return false;
